Add 'join-item' class to Btn component
In Btn.php, a 'join-item' class was added to handle grouping of buttons. This was done by checking if 'group' attribute exists in 'data' and then added to the 'attributes' array.
Additionally, the 'group' attribute was also removed from 'data' array to prevent further propagation since it serves the purpose only in Btn component.