"Refactor LazyComponent and add new Loading Component"
This commit contains a significant refactor of the LazyComponent and introduction of a new 'Loading' component. For the LazyComponent, adjustments were made to the handling of merging attributes and to how 'smart attributes' are added, increasing readability and organization of the code. Functionality for size handling was also added. The newly added 'Loading' Component includes support for various types of loading animations and handles size and type based on provided attributes. The changes are aimed at enhancing flexibility, improving code maintainability, and enriching the features of our component library.