Change name from 'lazy-component' to 'lazy-ui'
The name of the component has been changed from 'lazy-component' to 'lazy-ui'. This change was made across all the documentation, code and links to ensure uniformity and accurately reflect the purpose and usage of the component, which is to provide UI tools and features in a 'lazy' (i.e., easy-to-use and efficient) manner.