Add notify macros to LazyUiServiceProvider class
The Livewire Component class has been used to define two new macros: notify and notifyFlash in the LazyUiServiceProvider class. These macros will help in managing notifications in the application more efficiently. Now, it's easy to dispatch a notification or flash a notification to the session using these Livewire Component macros.