Adjust richtext and delete button functionality
In 'richtext.blade.php', replaced the x-lazy-richtext class with 'w-' class to correct the styling. In 'delete.blade.php', the href property is now dynamically filled, eliminating the previously hardcoded '#'. Also, upon successful deletion, the page will be redirected to a response data href or reloaded. The console.log response statement is removed for code cleanliness.