Add Quill editor along with toast notifications enhancement
Added Quill editor feature by implementing it in 'components/quill.js' to provide a rich text editor for improving the user input experience. This feature is imported in the 'lazy.js' file for general access. In parallel, an enhancement was applied to the toast notification. Now they are using `window.Alpine.store` in 'components/toast/index.js' instead of `Alpine.store` so that these notifications can be accessed globally.