Add 'axios' and default setup for 'lazy.scss' and 'lazy.js'
Added 'axios' to the project dependencies for handling HTTP requests. Altered the LazyInstallCommand.php to ensure that 'lazy.scss' and 'lazy.js' files are copied during the set up, if they do not already exists in the 'resources' directory. This is introduced to provide a preconfigured setup of SCSS and JS, to the users. New stub files for 'lazy.scss' and 'lazy.js' are included to be used as default files. This modification will ensure a smoother setup experience for the users.