Add Menu component to LazyUI
Added a new Menu component to the LazyUI library. The Menu.php file was added in the Components directory, and its corresponding view was created in resources/views/menu.blade.php. Also, the LazyUiServiceProvider.php file was modified to include this new component in the list. This new Menu component will provide a convenient way for developers to add menus in their UI, with flexible customization options.